remove array_distance, implement cosine_distance

+++ crates/runtime/src/embeddings/vector_search.rs
@@ -225,8 +225,8 @@ pub fn table_to_matches(
         .distances
         .iter()
         .flat_map(|v| {
-            if let Some(col) = v.as_any().downcast_ref::<arrow::array::Float32Array>() {
-                col.iter().collect::<Vec<Option<f32>>>()
+            if let Some(col) = v.as_any().downcast_ref::<arrow::array::Float64Array>() {
+                col.iter().collect::<Vec<Option<f64>>>()
             } else {
                 vec![]
             }
@@ -250,7 +250,7 @@ pub fn table_to_matches(
 
             Ok(Match {
                 value: value.clone(),
-                score: 1.0 / f64::from(*distance),
+                score: 1.0 / *distance,
                 dataset: tbl.to_string(),
                 primary_key: pks.get(i).cloned().unwrap_or_default(),
                 metadata: add_cols.get(i).cloned().unwrap_or_default(),
